refactor note test steps
