+++ features/steps/notes.js
@@ -16,34 +16,10 @@
    along with Ezra Bible App. See the file LICENSE.
    If not, see <http://www.gnu.org/licenses/>. */
 
-const { Given, When, Then, defineParameterType } = require("cucumber");
+const { Given, When, Then } = require("cucumber");
 const { expect } = require("chai");
-
-defineParameterType({
-  name: 'display_option',
-  regexp: /tags|notes|indicators|xrefs|footnotes/,
-  transformer: s => {
-    switch (s) {
-      case 'tags':
-        return '#tags-switch';
-      case 'notes':
-        return '#verse-notes-switch';
-      case 'indicators':
-        return '#user-data-indicators-switch';
-      case 'xrefs':
-        return '#xrefs-switch';
-      case 'footnotes':
-        return '#footnotes-switch';
-    }
-  }
-});
-
-defineParameterType({
-  name: 'state',
-  regexp: /displayed|hidden/,
-  type: 'boolean',
-  transformer: s => s == 'displayed'
-});
+const dbHelper = require("./helpers/db_helper.js");
+const uiHelper = require("./helpers/ui_helper.js");
 
 
 Given('I have {display_option} {state}', { timeout: 40 * 1000 }, async function (displayOptionId, state) {
@@ -61,7 +37,7 @@ Given('I have {display_option} {state}', { timeout: 40 * 1000 }, async function 
 });
 
 Given('I click on {string} note', async function (verseReference) {
-  var verseBox = await getVerseBox(verseReference);
+  var verseBox = await uiHelper.getVerseBox(verseReference);
   var classes = await verseBox.getAttribute('class');
 
   if(classes.split(' ').includes('book-notes')) {
@@ -75,7 +51,7 @@ Given('I click on {string} note', async function (verseReference) {
 });
 
 Given('I click on note indicator for the verse {string}', async function (verseReference) {
-  var verseBox = await getVerseBox(verseReference);
+  var verseBox = await uiHelper.getVerseBox(verseReference);
   var noteIndicator = await verseBox.$('.notes-info');
   noteIndicator.click()
   await spectronHelper.sleep(200);
@@ -99,13 +75,13 @@ When('I click note {string} button', async function (buttonClass) {
 Then('the note assigned to {string} in the database starts with text {string}', async function (verseReference, startText) {
   await global.spectronHelper.initDatabase();
 
-  var note = await global.models.Note.findByVerseReferenceId(await getDbVerseReferenceId(verseReference));
+  var note = await global.models.Note.findByVerseReferenceId(await dbHelper.getDbVerseReferenceId(verseReference));
 
   expect(note.text.startsWith(startText)).to.be.true;
 });
 
 Then('the note assigned to {string} has {string} text {string}', async function (verseReference, tag, text) {
-  var verseBox = await getVerseBox(verseReference);
+  var verseBox = await uiHelper.getVerseBox(verseReference);
   var verseNotesText = await verseBox.$('.verse-notes-text');
 
   if(!(await verseNotesText.isDisplayed())) {
@@ -119,50 +95,10 @@ Then('the note assigned to {string} has {string} text {string}', async function 
 });
 
 Then('the note assigned to {string} has {int} list items', async function (verseReference, num) {
-  var verseBox = await getVerseBox(verseReference);
+  var verseBox = await uiHelper.getVerseBox(verseReference);
   var verseNotesText = await verseBox.$('.verse-notes-text');
   var liElements = await verseNotesText.$$('li');
 
   expect(liElements.length).to.equal(num);
 });
 
-async function splitVerseReference(verseReference, translation = 'KJV') {
-  var [book, verseReferenceString] = verseReference.split(' ');
-  var bookId = global.spectronHelper.getBookShortTitle(book);
-
-  var verseReferenceHelper = await global.spectronHelper.getVerseReferenceHelper();
-  var absoluteVerseNumber = await verseReferenceHelper.referenceStringToAbsoluteVerseNr(translation, bookId, verseReferenceString);
-
-  return {
-    bookId,
-    absoluteVerseNumber
-  }
-}
-
-async function getVerseBox(verseReference) {
-  var verseListTabs = await global.app.client.$('#verse-list-tabs-1');
-  return await verseListTabs.$(await getVerseBoxSelector(verseReference));
-}
-
-async function getVerseBoxSelector(verseReference) {
-  var { absoluteVerseNumber } = await splitVerseReference(verseReference);
-  return `.verse-nr-${absoluteVerseNumber}`;
-}
-
-async function getDbVerseReferenceId(verseReference) {
-  var { bookId, absoluteVerseNumber } = await splitVerseReference(verseReference);
-
-  var dbBibleBook = await global.models.BibleBook.findOne({ where: { shortTitle: bookId } });
-  var dbVerseReference = await global.models.VerseReference.findOne({
-    where: {
-      bibleBookId: dbBibleBook.id,
-      absoluteVerseNrEng: absoluteVerseNumber
-    }
-  });
-
-  var allVerseReferences = await global.models.VerseReference.findAll();
-
-  expect(dbVerseReference, `Could not find a db verse reference for the given book (${dbBibleBook.id}) and absoluteVerseNr (${absoluteVerseNumber}). Total # of verse references: ${allVerseReferences.length}`).to.not.be.null;
-
-  return dbVerseReference.id;
-}
